% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticNNCBKMR.R
\name{find_ordered_nn_wvec}
\alias{find_ordered_nn_wvec}
\title{Find ordered nearest neighbors with weighted locations}
\usage{
find_ordered_nn_wvec(locs, w, m, wthres = 1e-04)
}
\arguments{
\item{locs}{An n x p matrix of locations (assumed to be ordered)}

\item{w}{A vector of length p of weights for each dimension}

\item{m}{The number of nearest neighbors to find}

\item{wthres}{A threshold below which weights are considered negligible}
}
\value{
An n x m matrix of nearest neighbor indices (1-based)
}
\description{
Find ordered nearest neighbors with weighted locations
}
\examples{
\dontrun{
locs <- matrix(runif(100), nrow=20)  # 20 points in 5D
w <- runif(5)  # weights for each dimension
m <- 3         # number of neighbors
nn_ind <- find_ordered_nn_wvec(locs, w, m)
}
}
