% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel_cross_RBF_rcpp_openmp}
\alias{kernel_cross_RBF_rcpp_openmp}
\title{Compute Cross-Covariance RBF kernel matrix between Z1 and Z2 using OpenMP}
\usage{
kernel_cross_RBF_rcpp_openmp(Z1, Z2, w)
}
\arguments{
\item{Z1}{An N1 x p matrix (can be 1 x p)}

\item{Z2}{An N2 x p matrix}

\item{w}{A vector of length p containing weights for each dimension}
}
\value{
An N1 x N2 RBF kernel matrix
}
\description{
Compute Cross-Covariance RBF kernel matrix between Z1 and Z2 using OpenMP
}
