% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel_mat_RBF_rcpp_openmp}
\alias{kernel_mat_RBF_rcpp_openmp}
\title{Compute RBF kernel matrix with corresponding weights using OpenMP}
\usage{
kernel_mat_RBF_rcpp_openmp(Z, w)
}
\arguments{
\item{Z}{An N x p matrix where each row is a data point.}

\item{w}{A vector of length p containing weights for each dimension.}
}
\value{
An N x N RBF kernel matrix.
}
\description{
This function computes the RBF kernel matrix for a given data matrix Z
and a vector of weights w, utilizing OpenMP for parallelization.
}
\examples{
\dontrun{
Z <- matrix(rnorm(100), nrow=10)  # 10 data points in 10D
w <- runif(10)                    # weights for each dimension
K <- kernel_mat_RBF_rcpp_openmp(Z, w)
}
}
