% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lCBpdf_bernoulli_no_X}
\alias{lCBpdf_bernoulli_no_X}
\title{Compute the log-density of a copula-based Bernoulli model without covariates
This function computes the log-density of a copula-based Bernoulli model
with only an intercept term (no covariates).}
\usage{
lCBpdf_bernoulli_no_X(y, beta0, Ut)
}
\arguments{
\item{y}{A vector of binary response variables (0s and 1s)}

\item{beta0}{The intercept parameter}

\item{Ut}{Either a matrix or a list containing the necessary components for the copula}
}
\value{
The log-density value of the copula-based Bernoulli model
}
\description{
Compute the log-density of a copula-based Bernoulli model without covariates
This function computes the log-density of a copula-based Bernoulli model
with only an intercept term (no covariates).
}
\examples{
\dontrun{
y <- c(0, 1, 1, 0, 1)  # Binary response vector
beta0 <- 0.5  # Intercept
Ut <- matrix(...)  # Some matrix or list for copula, e.g., from NNGP by running build_AD_rcpp
log_density <- lCBpdf_bernoulli_no_X(y, beta0, Ut)
}
}
