% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rnngp_loglik_tauonly_use_AD}
\alias{rnngp_loglik_tauonly_use_AD}
\title{Same as rnngp_loglik_tauonly but also return Ut as a list such that U.ImA = I-A
and U.D_inv_sqrt = D^{-1/2}}
\usage{
rnngp_loglik_tauonly_use_AD(x, coords, w, tau, neighbor_matrix, nugget = 1e-08)
}
\description{
Same as rnngp_loglik_tauonly but also return Ut as a list such that U.ImA = I-A
and U.D_inv_sqrt = D^{-1/2}
}
